made get all, get one and post on replacementOrder with dto returning the replacement order with orderItems

+++ src/main/java/kea/exercises/xpbowlingbackendkyll/model/stock/OrderItem.java
@@ -14,13 +14,13 @@ public class OrderItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
-    private double amountToOrder;
+    private int amountToOrder;
     @ManyToOne
     private StockItem stockItem;
     @ManyToOne
     private ReplacementOrder replacementOrder;
 
-    public OrderItem(double amountToOrder, StockItem stockItem, ReplacementOrder replacementOrder) {
+    public OrderItem(int amountToOrder, StockItem stockItem, ReplacementOrder replacementOrder) {
         this.amountToOrder = amountToOrder;
         this.stockItem = stockItem;
         this.replacementOrder = replacementOrder;
